MacroEditor: Clean up MacroStepEditor
Lift out EditInterval and EditWait, and remove unused imports.
